Move Inspector\DataCollector to DataCollector\EasyAdminDataCollector
